requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
pandas
requests
lxml
xlrd
dataframe_image
pyodbc
bs4
pretty-html-table
plotly
psutil
kaleido
matplotlib